Set the "non-coalesced" flag on OSX mouse events.
I don't see a thorough explanation of what this means in Apple's
documentation, but the "real" system appears to be setting it, and
I certainly don't think I'm doing anything that would count as
"coalescing", so...on it goes.

Note that it's only set on mouse events; setting in on keyboar
events appeared to screw up some thing (specific example: in iTerm2,
option-{b,f} inserted non-ascii charaters instead of moving the
insertion point).