More natural switch-by-mouse.
Now the pointer moves to the intuitive location on the switched-to
node when switching by mouse instead of wherever it last happened
to be (e.g. switching by tapping halfway up the left edge of the
current node's screen puts the pointer halfway up the right edge of
the screen on the next node).