Move edge-detection to master.
Remotes now respond to each MOVEREL they receive with a MOUSEPOS
message informing the master of the new position of their mouse
pointer after performing the move.  The READY message informs the
master of the remote's screen dimensions; in combination with the
position information from MOUSEPOS messages, the master now does
all the edge-event tracking itself.

This obviously comes at the expense of increased network traffic
(roughly ~2x during mouse movements), but it's still pretty small
potatoes, and worth it for the benefits of keeping complexity
out of the remotes and in the master.