Fix pointer lag when switching back to OSX master.
For some reason using CGWarpMouseCursorPosition() seems to cause a
~1/4 second or so period where the mouse motion events delivered by
a Quartz event tap all have the same position, which manifests
itself as a slight but annoying lag in cursor response when mouse-
switching back to an OSX master from a remote.

Here we work around this by adding a flag to ungrab_inputs() to
indicate whether or not the mouse position saved by grab_inputs()
should be restored, which we don't need when mouse-switching because
edgeswitch_reposition() is going to explicitly put it somewhere else
anyway, so it's only used when switching via hotkey (or refocusing
the master after a remote failure), where the lag is much less likely
to be noticeable (though I'm pretty sure it's still present).