x11: fix time passed to XSetSelectionOwner().
Previously we'd been passing in the manually (and certainly lossily)
maintained last_xevent_time, which usually worked but in some
circumstances would fail to take selection ownership (due to that
timestamp being too old).  After a bit of RTFM it appears we can just
pass the constant CurrentTime to mean "now", which is probably what we
wanted all along.

(Also provide a bit more information in the error message if it fails,
though that should hopefully stop happening now.)