x11: test XPending() to retrieve & handle X events, not just select().
X11 remotes can end up with pending (yet to be retrieved) events
queued in the Xlib buffer but a socket that would block if read from,
so select() doesn't notice anything to be done.  So as an additional
trigger for process_events() we now also check XPending() to cover
this case.

I'm not entirely certain why I'm only noticing this now.