Hack around a edge-state maintenance problem.
This arises with the new pointer-barrier-based edge event detection,
because when we do an edge-event-based switch to another node and then
back, we don't get a BarrierLeave event (because between a possible
mouse grab & reposition when switching away from the master and a
definite reposition when switching back regardless of node type the
"held at a barrier" state apparently gets dropped).  The old
mouse-position-tracking code, ugly as it was, got this right
"naturally".