x11: replace screen_center with grabbed_restpos
When we've got the mouse pointer grabbed, instead of keeping it at the
global center of the (aggregate) display, we now keep it at the center
of the largest single screen.  This is basically just a slight hack to
sidestep unfortunate interactions with xdpb -- with two equal-sized
displays, the global center is on the boundary between them, which
happens to be right where xdpb positions its own pointer barriers,
which interferes with the mouse events we're watching for to send
MOUSEMOVE events to the focused remote (which ends up causing some
fairly confusing behavior).  With grabbed_restpos a safe distance from
any screen edges, we avoid the problem.

(There's of course no guarantee there isn't a pointer barrier in the
middle of the largest screen either; in theory it should perhaps be
configurable, but I'm guessing this should do for common cases.)