Automate scheduled-call argument freeing
Callers are now no longer responsible for explicitly freeing the
opaque argument passed to schedule_call(); they instead pass an
optional function pointer to be called on 'arg' when the scheduled
call has either occurred or been canceled.  In addition to the general
ergonomic improvement, this resolves a minor memory leak that could
occur on an exit via hotkey with a remote focused with arguments for
scheduled brightness changes not getting freed when those scheduled
calls were then canceled.