Add -e/-d command-line flags for enabling/disabling remotes
This makes it easier to select exactly which remotes are in active use
by changing command-line flags instead of having to (say) comment or
uncomment things in a config file -- you can instead just leave a
collection of remotes defined and select the ones you want to use when
you start enthrall.

If only -e flags are used, other remotes are disabled by default, and
if only -d flags are used (or neither -d nor -e flags) other remotes
are enabled by default.  If both -d and -e flags are used, the default
enable state is determined by whichever one of the two is last in the
command line arguments.  Likewise, if the same remote is specified by
both an -e flag and a -d flag, whichever is later in the command line
will be in effect.